Modified I/O routines to accept 1-base row/columns
Modified input routines so that inputs for rows and columns are taken as 1-9, instead of the 0-8 that is used to process internally. Most of this is done in the utility __procCell method, which reduces its first two inputs by 1 to implement the change...remaining arguments, which are used for other identifiers, remain unchanged. Most methods that accept single items also accept arguments positionally without using procCells. This is important for cases like setEven, which passes its arguments to setEven positionally, thus not passing through procCells twice.